+++ src/Text/Show/Text/Data/String/UTF8.hs
@@ -10,7 +10,7 @@ Portability: GHC
 
 Monomorphic 'Show' function for 'UTF8' strings.
 
-/Since: 0.1/
+/Since: 0.2/
 -}
 module Text.Show.Text.Data.String.UTF8 (showbUTF8, toBuilder, toBuilderGeneric) where
 
@@ -34,18 +34,24 @@ import           Text.Show.Text.Utils ((<>), s)
 #include "inline.h"
 
 -- | Convert a 'UTF8' string to a 'Builder'.
+-- 
+-- /Since: 0.2/
 showbUTF8 :: UTF8Bytes string index => UTF8 string -> Builder
 showbUTF8 = showbBuilder . toBuilder
 {-# INLINE showbUTF8 #-}
 
 -- | Convert a 'UTF8'-encoded string to a 'Builder'.
 -- Invalid characters are replaced with '\xFFFD'.
+-- 
+-- /Since: 0.2/
 toBuilder :: UTF8Bytes string index => UTF8 string -> Builder
 toBuilder = toBuilderGeneric . toRep
 {-# INLINE toBuilder #-}
 
 -- | Convert a 'UTF8'-encoded bytestring to a 'Builder'.
 -- Invalid characters are replaced with '\xFFFD'.
+-- 
+-- /Since: 0.2/
 toBuilderGeneric :: UTF8Bytes bytestring index => bytestring -> Builder
 toBuilderGeneric = G.foldr ((<>) . s) mempty
 {-# SPECIALIZE toBuilderGeneric :: BS.ByteString -> Builder #-}