+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2
-* Added instances for the `binary`, `hoopl`, and `utf8-string` libraries
+* Exposed `showbKindPrec` and `showbPredPrec` with later versions of Template Haskell
+* Added instances for the `binary`, `hoopl`, `terminfo`, and `utf8-string` libraries
 
 # 0.1
 * Initial commit